Move MmCtx into its own module
Also avoid C type conflicts by using libc C types everywhere.

Also bindgen upgrade.
