+++ mm2src/common/Cargo.toml
@@ -30,7 +30,7 @@ tokio-core = "0.1"
 unwrap = "1.2"
 
 [build-dependencies]
-bindgen = "0.40"
+bindgen = "0.42"
 cc = "1.0"
 duct = "0.11"
 gstuff = "0.5"
