+++ mm2src/etomic/src/etomiccurl.rs
@@ -21,15 +21,15 @@
 
 #![allow(non_camel_case_types)]
 
-use common::{ post_json, fetch_json, MmArc };
+use common::{ post_json, fetch_json };
+use common::mm_ctx::MmArc;
 use common::log::StatusHandle;
 use common::etomiclib::*;
-use std::os::raw::{ c_char };
 use etomic::EthClient;
 use std::ffi::{ CStr, CString };
 use ethereum_types::{ U256, H256 };
 use hex;
-use libc;
+use libc::{self, c_char};
 use serde_json;
 use std::str::FromStr;
 use web3::{ Transport };
