+++ mm2src/rpc/commands.rs
@@ -17,7 +17,8 @@
 //  rpc_commands.rs
 //  marketmaker
 //
-use common::{lp, rpc_response, rpc_err_response, HyRes, MmArc, MM_VERSION};
+use common::{lp, rpc_response, rpc_err_response, HyRes, MM_VERSION};
+use common::mm_ctx::MmArc;
 use etomiccurl::get_gas_price_from_station;
 use ordermatch::{AutoBuyInput, lp_auto_buy};
 use serde_json::{self as json, Value as Json};
